some refactoring progress
