+++ src/mlm_form/templates.py
@@ -80,7 +80,8 @@ def trueFalseRadioTemplate(label, name, error_msg=None):
         style=f'{control_container_style} margin-bottom: 15px;'
     )
 
-def modelInputTemplate(label, name, error_msg=None):
+def modelInputTemplate(label, name, form_data, error_msg=None):
+    stats= update_statistics_by_norm(form_data.get('mlm_input_norm_type'))
     return Div(
         labelDecoratorTemplate(H4(label, style="margin-left: -30px;"), name in model_required_keys),
         Div(
@@ -97,14 +98,10 @@ def modelInputTemplate(label, name, error_msg=None):
         selectEnumTemplate("Input Data Type", datatypes,
             f"{name}_data_type", error_msg=None, canValidateInline=False),
         trueFalseRadioTemplate("Normalize Each Channel By Statistics", f"{name}_norm_by_channel", error_msg=None),
-        ## TODO this isn't working route not found
-        Div(
         selectEnumTemplate("Normalization Type", normalize_type_values,
-            f"{name}_norm_type", get="get_statistics", hx_target="#statistics", error_msg=None, canValidateInline=False),
-        Label("Statistics"),
-        Div(Div(id='statistics')),
-        ),
-
+            f"{name}_norm_type", error_msg=None, canValidateInline=False),
+        ## TODO this isn't working route not found
+        stats,
         selectEnumTemplate("Resize Type", resize_type_values,
             f"{name}_resize_type", error_msg=None, canValidateInline=False),
         Div(f'{error_msg}', style='color: red;') if error_msg else None,
@@ -149,7 +146,7 @@ def labelDecoratorTemplate(label, isRequired):
 
 def outputTemplate():
     return Div(
-        Div(id="result", style="position: fixed; right: 50px; width: 500px; height: calc(100vh - 250px); overflow: auto;"),
+        Div(json, id="result", style="position: fixed; right: 50px; width: 500px; height: calc(100vh - 250px); overflow: auto;"),
         style="position: relative;"
     )
 
@@ -158,3 +155,47 @@ def prettyJsonTemplate(obj):
 
 def error_template(msg):
     return  Div(msg, style='color: red; white-space: pre-wrap; margin-left: 10px; margin-bottom: 15px; text-indent: -10px;')
+
+def update_statistics_by_norm(mlm_input_norm_type: str):
+    if mlm_input_norm_type == 'none':
+        return Div(name='statistics')
+    elif mlm_input_norm_type == 'z-score':
+        return  Div(Div(
+                Label("Mean Statistic (enter a single comma separated list of values)"),
+                Input(type="text", name=f"mlm_input_mean", 
+                    placeholder='''e.g. 1354.40546513, 1118.24399958, 1042.92983953, 947.62620298,
+                                            1199.47283961, 1999.79090914, 2369.22292565,
+                                            2296.82608323, 732.08340178, 12.11327804,
+                                            1819.01027855, 1118.92391149, 2594.14080798''',
+                                            style=text_input_style),
+            ),
+            Div(
+                Label("Std Statistic (enter a single comma separated list of values)"),
+                Input(type="text", name=f"mlm_input_std", 
+                    placeholder='''e.g. 245.71762908, 333.00778264, 395.09249139,
+                                        593.75055589,
+                                        566.4170017,
+                                        861.18399006,
+                                        1086.63139075,
+                                        1117.98170791,
+                                        404.91978886,
+                                        4.77584468,
+                                        1002.58768311,
+                                        761.30323499,
+                                        1231.58581042''',
+                                        style=text_input_style),
+
+            ), name='statistics')
+    elif mlm_input_norm_type == 'min-max':
+        return  Div(
+                Div(
+                Label("Min Statistic (enter a single comma separated list of values)"),
+                Input(type="text", name=f"mlm_input_min", 
+                    placeholder='''e.g. ....''', style=text_input_style)),
+                Div(
+                Label("Max Statistic (enter a single comma separated list of values)"),
+                Input(type="text", name=f"mlm_input_max", 
+                    placeholder='''e.g. ....''', style=text_input_style)),
+                name='statistics')
+    else:
+        return Div(name='statistics')
